# docs/swagger.yaml
basePath: /
definitions:
...
  request.CarRequest:
    properties:
      carId:
        type: string
    type: object
...
  response.Success:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
  description: This is the doumentation for drag racing game
  title: Drag Racing
  version: "2.0"
paths:
  /car/buy:
    post:
      consumes:
      - application/json
      description: Buy a car for a player
      parameters:
      - description: Access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Buy Car Request
        in: body
        name: carRequest
        required: true
        schema:
          $ref: '#/definitions/request.CarRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Car added to player successfully
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Success'
        "401":
          description: Unauthorised
          schema:
            $ref: '#/definitions/response.Success'
        "404":
          description: Car not found
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Success'
      security:
      - ApiKeyAuth: []
      summary: Buy Car
      tags:
      - Car
...
  /car/equip:
    put:
      consumes:
      - application/json
      description: Equip a car for a player
      parameters:
      - description: Access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Equip Car Request
        in: body
        name: equipRequest
        required: true
        schema:
          $ref: '#/definitions/request.CarRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Car equipped successfully
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Success'
        "401":
          description: Unauthorised
          schema:
            $ref: '#/definitions/response.Success'
      summary: Equip Car
      tags:
      - Car
  /car/get-all:
    get:
      consumes:
      - application/json
      description: Retrieve the list of all car
      produces:
      - application/json
      responses:
        "200":
          description: Garage list fetched successfully
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Get All Garage List
      tags:
      - Car
  /car/get-by-id:
    post:
      consumes:
      - application/json
      description: Retrieve car details, stats, and customizations by car ID
      parameters:
      - description: Get car by ID request
        in: body
        name: getReq
        required: true
        schema:
          $ref: '#/definitions/request.CarRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Data fetch success
          schema:
            $ref: '#/definitions/response.Success'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Get car by ID
      tags:
      - Car
  /car/repair:
    post:
      consumes:
      - application/json
      description: Repair a car by deducting the repair cost from the player's coins
        and updating the car's durability
      parameters:
      - description: Repair car request
        in: body
        name: repairCarRequest
        required: true
        schema:
          $ref: '#/definitions/request.CarRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Car sold successfully
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Success'
        "401":
          description: Unauthorised
          schema:
            $ref: '#/definitions/response.Success'
        "404":
          description: Car not found
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Success'
      summary: Repair car
      tags:
      - Car
  /car/sell:
    delete:
      consumes:
      - application/json
      description: Sell a car for a player
      parameters:
      - description: Access token
        in: header
        name: Authorization
        required: true
        type: string
      - description: Sell Car Request
        in: body
        name: sellCarRequest
        required: true
        schema:
          $ref: '#/definitions/request.CarRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Car sold successfully
          schema:
            $ref: '#/definitions/response.Success'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.Success'
        "401":
          description: Unauthorised
          schema:
            $ref: '#/definitions/response.Success'
        "404":
          description: Car not found
          schema:
            $ref: '#/definitions/response.Success'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Success'
      security:
      - ApiKeyAuth: []
      summary: Sell Car
      tags:
      - Car
...
swagger: "2.0"